Fix sort order for more complex status indicators
